Fix HTTPS contexts experiment to work as reverse proxy
This commit provides a (mostly correct) working representation of the
Nginx reverse proxy pattern that uses context roots (e.g. /www and /app)
to direct to different upstreams. The location rewrite needs to be
improved a little bit, largely because /www and /app won't reroute to
the index page, but /www/index.html and /app/index.html will.